requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-RESTX==1.3.0
psycopg2-binary==2.9.7
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
celery==5.3.4
redis==5.0.1
webdriver-manager==4.0.1
lxml==4.9.3
Werkzeug==2.3.7
gunicorn==21.2.0 